+++ src/ResourceManager/Network/Commands.Network.Test/ScenarioTests/ApplicationGatewayTests.ps1
@@ -12,6 +12,42 @@
 # limitations under the License.
 # ----------------------------------------------------------------------------------
 
+function Test-AvailableSslOptions
+{
+	$result = Get-AzureRmApplicationGatewayAvailableSslOptions
+	Assert-NotNull $result
+	Assert-NotNull $result.DefaultPolicy
+
+	$result = Get-AzureRmApplicationGatewaySslPredefinedPolicy
+	Assert-NotNull $result
+	Assert-True { $result.Count -gt 0 }
+	Assert-NotNull $result[0].MinProtocolVersion
+	Assert-True { $result[0].CipherSuites -gt 0 }
+
+	$result = Get-AzureRmApplicationGatewaySslPredefinedPolicy -Name AppGwSslPolicy20170401
+	Assert-NotNull $result
+	Assert-NotNull $result.MinProtocolVersion
+	Assert-True { $result.CipherSuites -gt 0 }
+}
+
+function Test-AvailableWafRuleSets
+{
+	$result = Get-AzureRmApplicationGatewayAvailableWafRuleSets
+
+	Assert-NotNull $result
+	Assert-NotNull $result.Value
+	Assert-True { $result.Value.Count -gt 0 }
+	Assert-NotNull $result.Value[0].Name
+	Assert-NotNull $result.Value[0].RuleSetType
+	Assert-NotNull $result.Value[0].RuleSetVersion
+	Assert-NotNull $result.Value[0].RuleGroups
+	Assert-True { $result.Value[0].RuleGroups.Count -gt 0 }
+	Assert-NotNull $result.Value[0].RuleGroups[0].RuleGroupName
+	Assert-NotNull $result.Value[0].RuleGroups[0].Rules
+	Assert-True { $result.Value[0].RuleGroups[0].Rules.Count -gt 0 }
+	Assert-NotNull $result.Value[0].RuleGroups[0].Rules[0].RuleId
+}
+
 <#
 .SYNOPSIS
 Application gateway tests
@@ -119,7 +155,9 @@ function Test-ApplicationGatewayCRUD
 
 		$sslPolicy = New-AzureRmApplicationGatewaySslPolicy -DisabledSslProtocols TLSv1_0, TLSv1_1
 
-		$firewallConfig = New-AzureRmApplicationGatewayWebApplicationFirewallConfiguration -Enabled $true -FirewallMode Prevention
+		$disabledRuleGroup1 = New-AzureRmApplicationGatewayFirewallDisabledRuleGroupConfig -RuleGroupName "crs_41_sql_injection_attacks" -Rules 981318,981320
+		$disabledRuleGroup2 = New-AzureRmApplicationGatewayFirewallDisabledRuleGroupConfig -RuleGroupName "crs_35_bad_robots"
+		$firewallConfig = New-AzureRmApplicationGatewayWebApplicationFirewallConfiguration -Enabled $true -FirewallMode Prevention -RuleSetType "OWASP" -RuleSetVersion "2.2.9" -DisabledRuleGroups $disabledRuleGroup1,$disabledRuleGroup2
 
 		# Create Application Gateway
 		$appgw = New-AzureRmApplicationGateway -Name $appgwName -ResourceGroupName $rgname -Location $location -Probes $probeHttps -BackendAddressPools $pool, $nicPool -BackendHttpSettingsCollection $poolSetting01,$poolSetting02 -FrontendIpConfigurations $fipconfig01, $fipconfig02  -GatewayIpConfigurations $gipconfig -FrontendPorts $fp01, $fp02 -HttpListeners $listener01, $listener02 -RequestRoutingRules $rule01, $rule02 -Sku $sku -SslPolicy $sslPolicy -AuthenticationCertificates $authcert01 -WebApplicationFirewallConfiguration $firewallConfig
@@ -130,6 +168,7 @@ function Test-ApplicationGatewayCRUD
 		Assert-AreEqual "Running" $getgw.OperationalState
 		Compare-ConnectionDraining $poolSetting01 $getgw.BackendHttpSettingsCollection[0]
 		Compare-ConnectionDraining $poolSetting02 $getgw.BackendHttpSettingsCollection[1]
+		Compare-WebApplicationFirewallConfiguration $firewallConfig $getgw.WebApplicationFirewallConfiguration
 
 		# Get Application Gateway backend health with expanded resource
 		$backendHealth = Get-AzureRmApplicationGatewayBackendHealth -Name $appgwName -ResourceGroupName $rgname -ExpandResource "backendhealth/applicationgatewayresource"
@@ -148,7 +187,7 @@ function Test-ApplicationGatewayCRUD
         $nic01 = $nic01 | Set-AzureRmNetworkInterface
         $nic02 = $nic02 | Set-AzureRmNetworkInterface
 
-		# Add probe, request timeout, multi-hosting, URL routing to an exisitng gateway
+		# Add probe, request timeout, multi-hosting, URL routing to an existing gateway
 		# Probe, request timeout, multi-site and URL routing are optional.
 		$probeName = Get-ResourceName
 		$frontendPort03Name = Get-ResourceName
@@ -191,8 +230,16 @@ function Test-ApplicationGatewayCRUD
 
 		# Modify WAF config and verify that it can be retrieved
 		$getgw = Set-AzureRmApplicationGatewayWebApplicationFirewallConfiguration -ApplicationGateway $getgw -Enabled $true -FirewallMode Detection
-		Set-AzureRmApplicationGateway -ApplicationGateway $getgw
-		$firewallConfig2 = Get-AzureRmApplicationGatewayWebApplicationFirewallConfiguration -ApplicationGateway $getgw
+		$firewallConfig2 = Get-AzureRmApplicationGatewayWebApplicationFirewallConfiguration -ApplicationGateway $getgw		
+
+		# Verify that default values got set
+		Assert-AreEqual "OWASP"  $firewallConfig2.RuleSetType
+		Assert-AreEqual "3.0"  $firewallConfig2.RuleSetVersion
+		Assert-AreEqual $null  $firewallConfig2.DisabledRuleGroups
+
+		$getgw = Set-AzureRmApplicationGateway -ApplicationGateway $getgw
+
+		Compare-WebApplicationFirewallConfiguration $firewallConfig2 $getgw.WebApplicationFirewallConfiguration
 
 		# Remove probe, request timeout, multi-site and URL routing from exiting gateway
 		# Probe, request timeout, multi-site, URL routing are optional
@@ -238,6 +285,135 @@ function Test-ApplicationGatewayCRUD
 	}
 }
 
+<#
+.SYNOPSIS
+Application gateway tests
+#>
+function Test-ApplicationGatewayCRUD2
+{
+	param 
+	( 
+		$basedir = ".\" 
+	) 
+
+	# Setup	
+
+	$rglocation = Get-ProviderLocation ResourceManagement
+	$resourceTypeParent = "Microsoft.Network/applicationgateways"
+	$location = Get-ProviderLocation $resourceTypeParent
+
+	$rgname = Get-ResourceGroupName
+	$appgwName = Get-ResourceName
+	$vnetName = Get-ResourceName
+	$gwSubnetName = Get-ResourceName
+	$nicSubnetName = Get-ResourceName
+	$publicIpName = Get-ResourceName
+	$gipconfigname = Get-ResourceName
+
+	$frontendPort01Name = Get-ResourceName
+	$frontendPort02Name = Get-ResourceName
+	$fipconfigName = Get-ResourceName
+	$listener01Name = Get-ResourceName
+	$listener02Name = Get-ResourceName
+
+	$poolName = Get-ResourceName
+	$poolSetting01Name = Get-ResourceName
+
+	$redirect01Name = Get-ResourceName
+	$redirect02Name = Get-ResourceName
+	$redirect03Name = Get-ResourceName
+	$rule01Name = Get-ResourceName
+	$rule02Name = Get-ResourceName
+
+	$probeHttpName = Get-ResourceName
+
+	try 
+	{
+		# Create the resource group
+		$resourceGroup = New-AzureRmResourceGroup -Name $rgname -Location $location -Tags @{ testtag = "APPGw tag"} 
+      
+		# Create the Virtual Network
+		$gwSubnet = New-AzureRmVirtualNetworkSubnetConfig -Name $gwSubnetName -AddressPrefix 10.0.0.0/24
+		$nicSubnet = New-AzureRmVirtualNetworkSubnetConfig  -Name $nicSubnetName -AddressPrefix 10.0.2.0/24
+		$vnet = New-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname -Location $location -AddressPrefix 10.0.0.0/16 -Subnet $gwSubnet, $nicSubnet
+		$vnet = Get-AzureRmvirtualNetwork -Name $vnetName -ResourceGroupName $rgname
+		$gwSubnet = Get-AzureRmVirtualNetworkSubnetConfig -Name $gwSubnetName -VirtualNetwork $vnet
+ 		$nicSubnet = Get-AzureRmVirtualNetworkSubnetConfig -Name $nicSubnetName -VirtualNetwork $vnet
+
+		# Create public ip
+		$publicip = New-AzureRmPublicIpAddress -ResourceGroupName $rgname -name $publicIpName -location $location -AllocationMethod Dynamic
+
+		# Create ip configuration
+		$gipconfig = New-AzureRmApplicationGatewayIPConfiguration -Name $gipconfigname -Subnet $gwSubnet
+
+		#frontend part
+		$fipconfig = New-AzureRmApplicationGatewayFrontendIPConfig -Name $fipconfigName -PublicIPAddress $publicip
+		$fp01 = New-AzureRmApplicationGatewayFrontendPort -Name $frontendPort01Name  -Port 80
+		$fp02 = New-AzureRmApplicationGatewayFrontendPort -Name $frontendPort02Name  -Port 81
+		$listener01 = New-AzureRmApplicationGatewayHttpListener -Name $listener01Name -Protocol Http -FrontendIPConfiguration $fipconfig -FrontendPort $fp01
+		$listener02 = New-AzureRmApplicationGatewayHttpListener -Name $listener02Name -Protocol Http -FrontendIPConfiguration $fipconfig -FrontendPort $fp02
+
+		# backend part
+		$pool = New-AzureRmApplicationGatewayBackendAddressPool -Name $poolName -BackendIPAddresses www.microsoft.com, www.bing.com
+		$match = New-AzureRmApplicationGatewayProbeHealthResponseMatch -Body "helloworld" -StatusCode "200-300","404"
+		$probeHttp = New-AzureRmApplicationGatewayProbeConfig -Name $probeHttpName -Protocol Http -HostName "probe.com" -Path "/path/path.htm" -Interval 89 -Timeout 88 -UnhealthyThreshold 8 -Match $match
+		$poolSetting01 = New-AzureRmApplicationGatewayBackendHttpSettings -Name $poolSetting01Name -Port 80 -Protocol Http -Probe $probeHttp -CookieBasedAffinity Enabled -ProbeEnabled -PickHostNameFromBackendAddress
+
+		# rule part
+		$redirect01 = New-AzureRmApplicationGatewayRedirectConfiguration -Name $redirect01Name -RedirectType Permanent -TargetListener $listener01
+
+		$rule01 = New-AzureRmApplicationGatewayRequestRoutingRule -Name $rule01Name -RuleType basic -BackendHttpSettings $poolSetting01 -HttpListener $listener01 -BackendAddressPool $pool
+		$rule02 = New-AzureRmApplicationGatewayRequestRoutingRule -Name $rule02Name -RuleType basic -HttpListener $listener02 -RedirectConfiguration $redirect01
+
+		$sku = New-AzureRmApplicationGatewaySku -Name Standard_Medium -Tier Standard -Capacity 2
+
+		# security part
+		$sslPolicy = New-AzureRmApplicationGatewaySslPolicy -PolicyType Custom -MinProtocolVersion TLSv1_1 -CipherSuite "TLS_ECDHE_ECDSA_WITH_AES_128_GCM_SHA256", "TLS_ECDHE_ECDSA_WITH_AES_256_GCM_SHA384", "TLS_ECDHE_RSA_WITH_AES_128_CBC_SHA", "TLS_RSA_WITH_AES_128_GCM_SHA256"
+
+		# Create Application Gateway
+		$appgw = New-AzureRmApplicationGateway -Name $appgwName -ResourceGroupName $rgname -Location $location -Probes $probeHttp -BackendAddressPools $pool -BackendHttpSettingsCollection $poolSetting01 -FrontendIpConfigurations $fipconfig -GatewayIpConfigurations $gipconfig -FrontendPorts $fp01, $fp02 -HttpListeners $listener01, $listener02 -RedirectConfiguration $redirect01 -RequestRoutingRules $rule01, $rule02 -Sku $sku -SslPolicy $sslPolicy 
+
+		# Check get/set/remove for RedirectConfiguration
+		$redirect02 = Get-AzureRmApplicationGatewayRedirectConfiguration -ApplicationGateway $appgw -Name $redirect01Name
+		Assert-AreEqual $redirect01.TargetListenerId $redirect02.TargetListenerId
+		$getgw = Set-AzureRmApplicationGatewayRedirectConfiguration -ApplicationGateway $appgw -Name $redirect01Name -RedirectType Permanent -TargetUrl "https://www.bing.com"
+
+		$getgw = Add-AzureRmApplicationGatewayRedirectConfiguration -ApplicationGateway $getgw -Name $redirect03Name -RedirectType Permanent -TargetListener $listener01 -IncludePath $true
+		$getgw = Remove-AzureRmApplicationGatewayRedirectConfiguration -ApplicationGateway $getgw -Name $redirect03Name
+
+		# Get for SslPolicy
+		$sslPolicy01 = Get-AzureRmApplicationGatewaySslPolicy -ApplicationGateway $getgw
+		Assert-AreEqual $sslPolicy.MinProtocolVersion $sslPolicy01.MinProtocolVersion
+
+		# Set for sslPolicy
+		$getgw = Set-AzureRmApplicationGatewaySslPolicy -ApplicationGateway $getgw -PolicyType Predefined -PolicyName AppGwSslPolicy20170401
+
+		# Get Match
+		$probeHttp01 = Get-AzureRmApplicationGatewayProbeConfig -ApplicationGateway $getgw -Name $probeHttpName
+		Assert-AreEqual $probeHttp.Match.Body $probeHttp01.Match.Body
+
+		# Get Application Gateway
+		$getgw = Get-AzureRmApplicationGateway -Name $appgwName -ResourceGroupName $rgname
+
+		# Modify existing application gateway with new configuration
+		$getgw = Set-AzureRmApplicationGateway -ApplicationGateway $getgw
+
+		Assert-AreEqual "Running" $getgw.OperationalState
+
+		# Stop Application Gateway
+		$getgw = Stop-AzureRmApplicationGateway -ApplicationGateway $getgw
+
+		Assert-AreEqual "Stopped" $getgw.OperationalState
+ 
+		# Delete Application Gateway
+		Remove-AzureRmApplicationGateway -Name $appgwName -ResourceGroupName $rgname -Force
+	}
+	finally
+	{
+		# Cleanup
+		Clean-ResourceGroup $rgname
+	}
+}
 
 <#
 .SYNOPSIS
@@ -258,6 +434,67 @@ function Compare-ConnectionDraining($expected, $actual)
 	else 
 	{
 		Assert-Null $actualConnectionDraining
+	}
+}
+
+<#
+.SYNOPSIS
+Compare web application firewall configuration
+#>
+function Compare-WebApplicationFirewallConfiguration($expected, $actual) 
+{
+	if($expected) 
+	{
+		Assert-NotNull $actual
+		Assert-AreEqual $expected.Enabled $actual.Enabled
+		Assert-AreEqual $expected.FirewallMode $actual.FirewallMode
+		Assert-AreEqual $expected.RuleSetType $actual.RuleSetType
+		Assert-AreEqual $expected.RuleSetVersion $actual.RuleSetVersion
+
+		if($expected.DisabledRuleGroups) 
+		{
+			Assert-NotNull $actual.DisabledRuleGroups
+			Assert-AreEqual $expected.DisabledRuleGroups.Count $actual.DisabledRuleGroups.Count
+			for($i = 0; $i -lt $expected.DisabledRuleGroups.Count; $i++) 
+			{
+				Compare-DisabledRuleGroup $expected.DisabledRuleGroups[$i] $actual.DisabledRuleGroups[$i]
+			}
+		}
+		else
+		{
+			Assert-Null $actual.DisabledRuleGroups
+		}
+	}
+	else
+	{
+		Assert-Null $actual
+	}
+}
+
+<#
+.SYNOPSIS
+Compare disabled rule groups
+#>
+function Compare-DisabledRuleGroup($expected, $actual) 
+{
+	if($expected) 
+	{
+		Assert-NotNull $actual
+		Assert-AreEqual $expected.RuleGroupName $actual.RuleGroupName
+
+		if($expected.Rules) 
+		{
+			Assert-NotNull $actual.Rules
+			Assert-AreEqualArray $expected.Rules $actual.Rules
+		}
+		else
+		{
+			Assert-Null $actual.Rules
+		}
+	}
+	else
+	{
+		Assert-Null $actual
 	}
 }
 
@@ -284,4 +521,5 @@ function Compare-AzureRmApplicationGateway($expected, $actual)
 
 	Assert-AreEqual $expected.HttpListeners.Count $actual.HttpListeners.Count
 	Assert-AreEqual $expected.RequestRoutingRules.Count $actual.RequestRoutingRules.Count
+	Assert-AreEqual $expected.RedirectConfigurations.Count $actual.RedirectConfigurations.Count
 }