+++ src/ResourceManager/Network/ChangeLog.md
@@ -19,6 +19,79 @@
 -->
 ## Current Release
 
+## Version 4.2.1
+
+## Version 4.2.0
+* New-AzureRmIpsecPolicy: SALifeTimeSeconds and SADataSizeKilobytes are no longer mandatory parameters
+    - SALifeTimeSeconds defaults to 27000 seconds
+    - SADataSizeKilobytes defaults to 102400000 KB
+* Added support for custom cipher suite configuration using ssl policy and listing all ssl options api in Application Gateway
+    - Added optional parameter -PolicyType, -PolicyName, -MinProtocolVersion, -Ciphersuite
+        - Add-AzureRmApplicationGatewaySslPolicy
+        - New-AzureRmApplicationGatewaySslPolicy
+        - Set-AzureRmApplicationGatewaySslPolicy
+    - Added Get-AzureRmApplicationGatewayAvailableSslOptions (Alias: List-AzureRmApplicationGatewayAvailableSslOptions)
+    - Added Get-AzureRmApplicationGatewaySslPredefinedPolicy (Alias: List-AzureRmApplicationGatewaySslPredefinedPolicy)
+* Added redirect support in Application Gateway
+    - Added Add-AzureRmApplicationGatewayRedirectConfiguration
+    - Added Get-AzureRmApplicationGatewayRedirectConfiguration
+    - Added New-AzureRmApplicationGatewayRedirectConfiguration
+    - Added Remove-AzureRmApplicationGatewayRedirectConfiguration
+    - Added Set-AzureRmApplicationGatewayRedirectConfiguration
+    - Added optional parameter -RedirectConfiguration
+        - Add-AzureRmApplicationGatewayRequestRoutingRule
+        - New-AzureRmApplicationGatewayRequestRoutingRule
+        - Set-AzureRmApplicationGatewayRequestRoutingRule
+    - Added optional parameter -DefaultRedirectConfiguration
+        - Add-AzureRmApplicationGatewayUrlPathMapConfig
+        - New-AzureRmApplicationGatewayUrlPathMapConfig
+        - Set-AzureRmApplicationGatewayUrlPathMapConfig
+    - Added optional parameter -RedirectConfiguration
+        - Add-AzureRmApplicationGatewayPathRuleConfig
+        - New-AzureRmApplicationGatewayPathRuleConfig
+        - Set-AzureRmApplicationGatewayPathRuleConfig
+    - Added optional parameter -RedirectConfigurations
+        - New-AzureRmApplicationGateway 
+        - Set-AzureRmApplicationGateway
+* Added support for azure websites in Application Gateway
+    - Added New-AzureRmApplicationGatewayProbeHealthResponseMatch
+    - Added optional parameters -PickHostNameFromBackendHttpSettings, -MinServers, -Match
+        - Add-AzureRmApplicationGatewayProbeConfig 
+        - New-AzureRmApplicationGatewayProbeConfig
+        - Set-AzureRmApplicationGatewayProbeConfig
+    - Added optional parameters -PickHostNameFromBackendAddress, -AffinityCookieName, -ProbeEnabled, -Path
+        - Add-AzureRmApplicationGatewayBackendHttpSettings
+        - New-AzureRmApplicationGatewayBackendHttpSettings
+        - Set-AzureRmApplicationGatewayBackendHttpSettings
+* Update Get-AzureRmPublicIPaddress to retrieve publicipaddress resources created via VM Scale Set
+* Added cmdlet to get virtual network current usage
+    - Get-AzureRmVirtualNetworkUsageList
+
+## Version 4.1.0
+* Get-AzureRmNetworkUsage: New cmdlet to show network usage and capacity details
+* Added new GatewaySku options for VirtualNetworkGateways
+    - VpnGw1, VpnGw2, VpnGw3 are the new Skus added for Vpn gateways
+* Set-AzureRmNetworkWatcherConfigFlowLog
+  * Fixed  help examples
+ 
+## Version 4.0.1
+
+## Version 4.0.0
+* Added Test-AzureRmNetworkWatcherConnectivity cmdlet
+    - Returns connectivity information for a specified source VM and a destination
+    - If connectivity between the source and destination cannot be established, the cmdlet returns details about the issue
+
+## Version 3.7.0
+* Added support for new web application firewall features to Application Gateways
+    - Added New-AzureRmApplicationGatewayFirewallDisabledRuleGroupConfig
+    - Added Get-AzureRmApplicationGatewayAvailableWafRuleSets (Alias: List-AzureRmApplicationGatewayAvailableWafRuleSets)
+    - Updated New-AzureRmApplicationGatewayWebApplicationFirewallConfiguration: Added parameter -RuleSetType -RuleSetVersion and -DisabledRuleGroups
+    - Updated Set-AzureRmApplicationGatewayWebApplicationFirewallConfiguration: Added parameter -RuleSetType -RuleSetVersion and -DisabledRuleGroups
+
+* Added support for IPSec policies to Virtual Network Gateway Connections
+	- Added New-AzureRmIpsecPolicy
+	- Updated New-AzureRmVirtualNetworkGatewayConnection: Added parameter -IpsecPolicies and -UsePolicyBasedTrafficSelectors
+
 ## Version 3.6.0
 * Added support for connection draining to Application Gateways
     - Added Get-AzureRmApplicationGatewayConnectionDraining