+++ src/ResourceManager/Scheduler/ChangeLog.md
@@ -19,6 +19,18 @@
 -->
 ## Current Release
 
+## Version 0.15.2
+
+## Version 0.15.1
+
+## Version 0.15.0
+
+## Version 0.14.1
+
+## Version 0.14.0
+
+## Version 0.13.0
+
 ## Version 0.12.0
 
 ## Version 0.11.4